refactor: ef op tests
